crm-app/src/workers/python/requirements.txt
# Core AI & ML
google-generativeai==0.3.2
numpy>=1.26.0
scipy>=1.11.0
scikit-learn>=1.3.0
pandas>=2.2.0

# Image Processing & OCR
Pillow>=10.2.0

# Excel & Google Sheets
openpyxl>=3.1.0
xlsxwriter>=3.2.0
gspread>=6.0.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0

# HTTP & API
requests>=2.31.0
facebook-business==19.0.0

# Data Storage
redis>=5.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0

# Vector DB (for RAG - Phase 8)
# chromadb>=0.4.0

# Visualization (optional)
matplotlib>=3.8.0
seaborn>=0.13.0
